+++ src/components/MainContainer/index.jsx
@@ -4,11 +4,11 @@ import { CreateCard } from "../CreateCard";
 import { Container, HeaderContent, MainCardContent } from "./styles";
 
 export function MainContainer(props) {
-    const { pokemons, loading, showCard, showInsertCard, CloseInsertCard, handleOpenNewDeleteModal } = props
+    const { pokemons, loading, showCard, showInsertCard, CloseInsertCard, handleOpenNewDeleteModal, notify } = props
 
     return (
         <Container>
-            <CreateCard showCard={showCard} CloseInsertCard={CloseInsertCard} />
+            <CreateCard showCard={showCard} CloseInsertCard={CloseInsertCard} notify={notify} />
             <HeaderContent>
                 <h2>Resultado de Busca</h2>
                 <button onClick={showInsertCard}>Novo Card</button>
@@ -21,6 +21,7 @@ export function MainContainer(props) {
                                 showInsertCard={showInsertCard}
                                 CloseInsertCard={CloseInsertCard}
                                 handleOpenNewDeleteModal={handleOpenNewDeleteModal}
+                                notify={notify}
                                 pokemon={pokemon}
                                 key={pokemon.name}
                             />
